NET-2: moved ResolveHostID to old hostnetwork
